vpn_api/requirements.txt
alembic~=1.11.1
fastapi~=0.100.1
pydantic~=2.1.1
python-dotenv~=1.0.0
SQLAlchemy~=2.0.19
psycopg2-binary
starlette~=0.27.0
passlib~=1.7.4
uvicorn~=0.23.1
requests~=2.31.0